dffdf

+++ ecos_chart.py
@@ -271,7 +271,7 @@ def transform_color(x):
 def fred_spread_chart(df1, df2):
     df2 = df2.dropna()
     df1.loc[:,"기준금리차"] = df1.iloc[:,0] - df2.iloc[-1,0]
-    df1.loc[:,'color'] = np.where(df1['변동'].astype(float) < 0, '#FFB8B1', '#E2F0CB')
+    df1.loc[:,'color'] = np.where(df1['변동'] < 0, '#FFB8B1', '#E2F0CB')
     # df1.loc[:,'color'] = df1['변동'].astype(float).apppy(lambda x: transform_color(x))
     df2.loc[:,'10Y2Ycolor'] = np.where(df2['금리차10Y2Y']<0, '#FFB8B1', '#E2F0CB')
     df2.loc[:,'10Y3Mcolor'] = np.where(df2['금리차10Y3M']<0, '#FFB8B1', '#E2F0CB')
